+++ src/oaklib/implementations/sqldb/sql_implementation.py
@@ -98,7 +98,7 @@ from oaklib.interfaces.basic_ontology_interface import (
     RELATIONSHIP_MAP,
     BasicOntologyInterface,
 )
-from oaklib.interfaces.class_enrichment_interface import ClassEnrichmentCalculationInterface
+from oaklib.interfaces.class_enrichment_calculation_interface import ClassEnrichmentCalculationInterface
 from oaklib.interfaces.differ_interface import DifferInterface
 from oaklib.interfaces.mapping_provider_interface import MappingProviderInterface
 from oaklib.interfaces.metadata_interface import MetadataInterface
