+++ tests/test_implementations/__init__.py
@@ -33,7 +33,7 @@ from oaklib.interfaces.association_provider_interface import (
     AssociationProviderInterface,
     associations_subjects,
 )
-from oaklib.interfaces.class_enrichment_interface import ClassEnrichmentCalculationInterface
+from oaklib.interfaces.class_enrichment_calculation_interface import ClassEnrichmentCalculationInterface
 from oaklib.interfaces.differ_interface import DifferInterface
 from oaklib.interfaces.obograph_interface import OboGraphInterface
 from oaklib.interfaces.patcher_interface import PatcherInterface
